pyproject.toml
[project]
name = "hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "openai>=2.24.0",
    "pymupdf>=1.27.1",
    "python-dotenv>=1.2.1",
]
